Rename router url
