+++ src/components/Header.vue
@@ -88,9 +88,9 @@ const  SIMPLE_MENU = [
         summary: 'header.bgpMonitor.summary',
       },
       {
-        entryName: 'header.gsoc24TracerouteVisualization.name',
-        routeName: 'gsoc24-traceroute-visualization',
-        summary: 'header.gsoc24TracerouteVisualization.summary',
+        entryName: 'header.tracerouteVisualization.name',
+        routeName: 'traceroute-visualization',
+        summary: 'header.tracerouteVisualization.summary',
       },
     ],
   },
